实现单机连接

+++ termux_init.sh
@@ -1,34 +1,43 @@
 #!/bin/sh
 
-pkg update
-
-# 安装必要环境
-pkg install -y python ndk-sysroot clang make libjpeg-turbo libxml2 libxslt termux-api git nodejs-lts
-
-# adb 安装
-cd ~
-git clone https://github.com/Magisk-Modules-Repo/adb-ndk.git
-cd ./adb-ndk/bin/
-mv -f adb.bin adb
-chmod +x ./*
-mv -f ./* $PREFIX/bin/
-cd ../..
-rm -rf adb-ndk/
+adb version
+if [ $? -ne 0 ]; then
+    # adb 安装
+    cd ~
+    git clone https://github.com/Magisk-Modules-Repo/adb-ndk.git
+    cd ./adb-ndk/bin/
+    mv -f adb.bin adb
+    chmod +x ./*
+    mv -f ./* $PREFIX/bin/
+    cd ../..
+    rm -rf adb-ndk/
 # adb 安装结束
+fi
 
-# 升级pip3
-python -m pip3 install --upgrade pip3
+python3 -c "import uiautomator2"
+if [ $? -ne 0 ]; then
+    pkg update
 
-# 安装U2
-pip3 install -U uiautomator2
+    # 安装必要环境
+    pkg install -y python ndk-sysroot clang make libjpeg-turbo libxml2 libxslt termux-api git nodejs-lts
 
-# 初始化一次
-python3 -m uiautomator2 init
+    # 升级pip3
+    python -m pip3 install --upgrade pip3
 
-#创建测试
-script='import uiautomator2 as u2
+    # 安装U2
+    pip3 install -U uiautomator2
+
+    # 初始化一次
+    python3 -m uiautomator2 init
+
+    #创建测试
+    script='
+import uiautomator2 as u2
 d = u2.connect("0.0.0.0")
-print(d.info)'
+print(d.info)
+    '
 
-echo "$script" >test.py
-python3 test.py
+    echo "$script" >test.py
+    python3 test.py
+
+fi
